SerializeAddon: move test util into seperate file
