+++ addons/xterm-addon-serialize/test/SerializeAddon.api.ts
@@ -18,12 +18,12 @@ const writeRawSync = (page: any, str: string): Promise<void> => writeSync(page, 
 
 const testNormalScreenEqual = async (page: any, str: string): Promise<void> => {
   await writeRawSync(page, str);
-  const originalBuffer = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+  const originalBuffer = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
   const result = await page.evaluate(`serializeAddon.serialize();`) as string;
   await page.evaluate(`term.reset();`);
   await writeRawSync(page, result);
-  const newBuffer = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+  const newBuffer = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
   // chai decides -0 and 0 are different number...
   // and firefox have a bug that output -0 for unknown reason
@@ -51,11 +51,11 @@ describe('SerializeAddon', () => {
 
   it('produce different output when we call test util with different text', async function(): Promise<any> {
     await writeRawSync(page, '12345');
-    const buffer1 = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+    const buffer1 = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
     await page.evaluate(`term.reset();`);
     await writeRawSync(page, '67890');
-    const buffer2 = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+    const buffer2 = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
     assert.throw(() => {
       assert.equal(JSON.stringify(buffer1), JSON.stringify(buffer2));
@@ -64,11 +64,11 @@ describe('SerializeAddon', () => {
 
   it('produce different output when we call test util with different line wrap', async function(): Promise<any> {
     await writeRawSync(page, '1234567890\r\n12345');
-    const buffer3 = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+    const buffer3 = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
     await page.evaluate(`term.reset();`);
     await writeRawSync(page, '1234567890n12345');
-    const buffer4 = await page.evaluate(`SerializeAddon._inspectBuffer(term.buffer.normal);`);
+    const buffer4 = await page.evaluate(`SerializeAddonTestUtil.inspectBuffer(term.buffer.normal);`);
 
     assert.throw(() => {
       assert.equal(JSON.stringify(buffer3), JSON.stringify(buffer4));
